Modularize tests for Oracle, MySQL, Postgres, SQL Server
